fix cloud tablet
