+++ fe/fe-core/src/main/java/org/apache/doris/catalog/LocalTablet.java
@@ -17,17 +17,28 @@
 
 package org.apache.doris.catalog;
 
+import org.apache.doris.clone.TabletSchedCtx;
+import org.apache.doris.clone.TabletSchedCtx.Priority;
+import org.apache.doris.common.Config;
+import org.apache.doris.common.FeConstants;
 import org.apache.doris.common.Pair;
+import org.apache.doris.system.Backend;
+import org.apache.doris.system.SystemInfoService;
 
 import com.google.gson.annotations.SerializedName;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import java.util.ArrayList;
 import java.util.Comparator;
+import java.util.Iterator;
+import java.util.List;
 
 public class LocalTablet extends Tablet {
     private static final Logger LOG = LogManager.getLogger(LocalTablet.class);
 
+    @SerializedName(value = "rs", alternate = {"replicas"})
+    private List<Replica> replicas;
     @SerializedName(value = "lastCheckTime")
     private long lastCheckTime;
 
@@ -63,6 +74,9 @@ public class LocalTablet extends Tablet {
 
     public LocalTablet(long tabletId) {
         super(tabletId);
+        if (this.replicas == null) {
+            this.replicas = new ArrayList<>();
+        }
         checkedVersion = -1L;
         isConsistent = true;
     }
@@ -160,5 +174,154 @@ public class LocalTablet extends Tablet {
     @Override
     public void setLastCheckTime(long lastCheckTime) {
         this.lastCheckTime = lastCheckTime;
+    }
+
+    private boolean isLatestReplicaAndDeleteOld(Replica newReplica) {
+        boolean delete = false;
+        boolean hasBackend = false;
+        long version = newReplica.getVersion();
+        Iterator<Replica> iterator = replicas.iterator();
+        while (iterator.hasNext()) {
+            Replica replica = iterator.next();
+            if (replica.getBackendIdWithoutException() == newReplica.getBackendIdWithoutException()) {
+                hasBackend = true;
+                if (replica.getVersion() <= version) {
+                    iterator.remove();
+                    delete = true;
+                }
+            }
+        }
+
+        return delete || !hasBackend;
+    }
+
+    @Override
+    public void addReplica(Replica replica, boolean isRestore) {
+        if (isLatestReplicaAndDeleteOld(replica)) {
+            replicas.add(replica);
+            if (!isRestore) {
+                Env.getCurrentInvertedIndex().addReplica(id, replica);
+            }
+        }
+    }
+
+    @Override
+    public List<Replica> getReplicas() {
+        return this.replicas;
+    }
+
+    @Override
+    public Replica getReplicaByBackendId(long backendId) {
+        for (Replica replica : replicas) {
+            if (replica.getBackendIdWithoutException() == backendId) {
+                return replica;
+            }
+        }
+        return null;
+    }
+
+    @Override
+    public boolean deleteReplica(Replica replica) {
+        if (replicas.contains(replica)) {
+            replicas.remove(replica);
+            Env.getCurrentInvertedIndex().deleteReplica(id, replica.getBackendIdWithoutException());
+            return true;
+        }
+        return false;
+    }
+
+    @Override
+    public boolean deleteReplicaByBackendId(long backendId) {
+        Iterator<Replica> iterator = replicas.iterator();
+        while (iterator.hasNext()) {
+            Replica replica = iterator.next();
+            if (replica.getBackendIdWithoutException() == backendId) {
+                iterator.remove();
+                Env.getCurrentInvertedIndex().deleteReplica(id, backendId);
+                return true;
+            }
+        }
+        return false;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj) {
+            return true;
+        }
+        if (!(obj instanceof LocalTablet)) {
+            return false;
+        }
+
+        LocalTablet tablet = (LocalTablet) obj;
+
+        if (replicas != tablet.replicas) {
+            if (replicas.size() != tablet.replicas.size()) {
+                return false;
+            }
+            int size = replicas.size();
+            for (int i = 0; i < size; i++) {
+                if (!tablet.replicas.contains(replicas.get(i))) {
+                    return false;
+                }
+            }
+        }
+        return id == tablet.id;
+    }
+
+    /**
+     * check if this tablet is ready to be repaired, based on priority.
+     * VERY_HIGH: repair immediately
+     * HIGH:    delay Config.tablet_repair_delay_factor_second * 1;
+     * NORMAL:  delay Config.tablet_repair_delay_factor_second * 2;
+     * LOW:     delay Config.tablet_repair_delay_factor_second * 3;
+     */
+    @Override
+    public boolean readyToBeRepaired(SystemInfoService infoService, TabletSchedCtx.Priority priority) {
+        if (FeConstants.runningUnitTest) {
+            return true;
+        }
+
+        if (priority == Priority.VERY_HIGH) {
+            return true;
+        }
+
+        boolean allBeAliveOrDecommissioned = true;
+        for (Replica replica : replicas) {
+            Backend backend = infoService.getBackend(replica.getBackendIdWithoutException());
+            if (backend == null || (!backend.isAlive() && !backend.isDecommissioned())) {
+                allBeAliveOrDecommissioned = false;
+                break;
+            }
+        }
+
+        if (allBeAliveOrDecommissioned) {
+            return true;
+        }
+
+        long currentTime = System.currentTimeMillis();
+
+        // first check, wait for next round
+        if (getLastStatusCheckTime() == -1) {
+            setLastStatusCheckTime(currentTime);
+            return false;
+        }
+
+        boolean ready = false;
+        switch (priority) {
+            case HIGH:
+                ready = currentTime - getLastStatusCheckTime() > Config.tablet_repair_delay_factor_second * 1000 * 1;
+                break;
+            case NORMAL:
+                ready = currentTime - getLastStatusCheckTime() > Config.tablet_repair_delay_factor_second * 1000 * 2;
+                break;
+            case LOW:
+                ready = currentTime - getLastStatusCheckTime() > Config.tablet_repair_delay_factor_second * 1000 * 3;
+                break;
+            default:
+                break;
+        }
+
+        return ready;
     }
 }
